requirements.txt
doctr
llama-cpp-python
